With propertyWrapper.
